% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoStepBootstrap.R
\name{TwoStepBootstrap}
\alias{TwoStepBootstrap}
\title{Estimation and Inference for Two-Step Estimators}
\usage{
TwoStepBootstrap(
  DatR,
  VarR,
  DatY,
  VarY,
  variogram.model,
  is.cov.misspecified,
  is.den.misspecified,
  plot.start.value = TRUE,
  cutoff,
  cutoff_u,
  start.value.method = 2,
  projected = FALSE
)
}
\arguments{
\item{DatR}{explanatory variable R, a spatial object, see \link[gstat:coordinates]{coordintes()}}

\item{VarR}{name of variable R}

\item{DatY}{outcome variable Y, a spatial object, see \link[gstat:coordinates]{coordintes()}}

\item{VarY}{name of variable Y}

\item{variogram.model}{variogram model type, e.g. "Exp", "Sph", "Gau", "Mat"}

\item{is.cov.misspecified}{logical; if TRUE, the covariance function is misspecified}

\item{is.den.misspecified}{logical; if TRUE, the density function is not Gaussian distribution.}

\item{plot.start.value}{logical, if TRUE, plot the variogram and the fitted variogram curve corresponding to starting values}

\item{start.value.method}{fitting method, see \link[gstat:fit.variogram]{fit.variogram()}}

\item{projected}{logical; if FALSE, data are assumed to be unprojected, meaning decimal longitude/latitude. For projected data, Euclidian distances are computed, for unprojected great circle distances(km) are computed.}
}
\value{

}
\description{
This function implements the esmation and inference for two-step estimators including Krig-and-regress(OLS), Krig-and-regress(GLS), two-step bootstrap.
}
\seealso{
\link{gstat}
}
