% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoStepBootstrap.R
\name{TwoStepBootstrap}
\alias{TwoStepBootstrap}
\title{Estimation and Inference for Two-Step Estimator including OLS, GLS, Krig-in Regress, and Bootstrap}
\usage{
TwoStepBootstrap(
  DatR,
  VarR,
  DatY,
  VarY,
  Variogram_Model,
  is_cov_misspecified,
  is_den_misspecified,
  Plot = TRUE,
  cutoff,
  cutoff_u,
  Start_Value_Method = 2,
  projected = FALSE
)
}
\arguments{
\item{DatR}{variable R, a spatial object, see function coordinates()}

\item{VarR}{name of variable R}

\item{DatY}{variable Y, a spatial object, see function coordinates()}

\item{VarY}{name of variable Y}

\item{Variogram_Model}{variogram model, output of}

\item{is_cov_misspecified}{logical; if TRUE, the covariance function is misspecified}

\item{is_den_misspecified}{logical; if TRUE, the density function is not Gaussian distribution.}

\item{Start_Value_Method}{fitting method, used by gstat, see function fit.variogram()}

\item{projected}{logical; if FALSE, data are assumed to be unprojected, meaning decimal longitude/latitude. For projected data, Euclidian distances are computed, for unprojected great circle distances(km)}

\item{Plot_Start_Value}{logical, plot the variogram and the fitted variogram curve corresponding to starting values}
}
\value{

}
\description{
This function implements the esmation and inference for two-step estimators.
}
